Rename curves folder to protocols
