+++ src/protocols/eddsa/signer.rs
